python-serial/testing/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "thermoflex"
version = "0.0.3"
authors = [{name = "Marvin Stephens", email = "marvin.stephens44@yahoo.com"}]
readme = "README_QUICK.MD"
dependencies = ['pyserial == 3.5','protobuf == 5.28.3']
description = "This is the repository for the thermoflex muscle by Delta Robotics"
keywords = ["thermoflex" , "Delta", "Delta Robotics"]
requires-python = ">= 3.8"


